During serialization memory allocation should honour allocator chunk size
Add a traits type for the memory pool/allocator and separate memory_pool
and allocator classes to make further customization (other parcelports)
cleaner.
