+++ hpx/runtime/parcelset/encode_parcels.hpp
@@ -12,6 +12,7 @@
 
 #include <hpx/runtime/parcelset/parcel_buffer.hpp>
 #include <hpx/util/high_resolution_timer.hpp>
+#include <hpx/traits/is_chunk_allocator.hpp>
 
 #include <boost/integer/endian.hpp>
 
@@ -86,6 +87,9 @@ namespace hpx
             // guard against serialization errors
             try {
                 try {
+                    // Get the chunk size from the allocator if it supports it
+                    size_t chunk_default =
+                        hpx::traits::default_chunk_size<typename Buffer::allocator_type>::call(buffer.data_.get_allocator());
                     // preallocate data
                     for (/**/; parcels_sent != parcels_size; ++parcels_sent)
                     {
@@ -93,8 +97,8 @@ namespace hpx
                             break;
                         arg_size += traits::get_type_size(ps[parcels_sent]);
                     }
-
-                    buffer.data_.reserve(arg_size);
+                    
+                    buffer.data_.reserve(std::max(chunk_default,arg_size));
 
                     // mark start of serialization
                     util::high_resolution_timer timer;
