+++ hpx/util/memory_chunk_pool.hpp
@@ -8,19 +8,33 @@
 
 #include <hpx/config.hpp>
 
+#include <hpx/traits/is_chunk_allocator.hpp>
 #include <hpx/util/memory_chunk.hpp>
-
+#include <hpx/util/memory_chunk_pool_allocator.hpp>
+//
 #include <cstdlib>
 #include <vector>
 
+// forward declare pool
 namespace hpx { namespace util
 {
-    namespace detail
-    {
-        template <typename T, typename Mutex = hpx::lcos::local::spinlock>
-        struct memory_chunk_pool_allocator;
-    }
+    template <typename Mutex>
+    struct memory_chunk_pool;
+}}
 
+// specialize chunk pool allocator traits for this memory_chunk_pool
+namespace hpx { namespace traits
+{
+    // if the chunk pool supplies fixed chunks of memory when the alloc
+    // is smaller than some threshold, then the pool must declare
+    // std::size_t chunk_size_
+    template <typename T, typename M>
+    struct is_chunk_allocator<util::detail::memory_chunk_pool_allocator<T,util::memory_chunk_pool<M>,M>>
+      : boost::mpl::true_ {};
+}}
+
+namespace hpx { namespace util
+{
     template <typename Mutex = hpx::lcos::local::spinlock>
     struct memory_chunk_pool : boost::noncopyable
     {
@@ -184,99 +198,6 @@ namespace hpx { namespace util
         std::size_t const backup_threshold_;
     };
 
-    namespace detail
-    {
-        template <typename T, typename Mutex>
-        struct memory_chunk_pool_allocator
-        {
-            typedef std::size_t size_type;
-            typedef std::ptrdiff_t difference_type;
-            typedef T* pointer;
-            typedef const T* const_pointer;
-            typedef T& reference;
-            typedef const T& const_reference;
-            typedef T value_type;
-
-            template <typename U>
-            struct rebind
-            {
-                typedef memory_chunk_pool_allocator<U, Mutex> other;
-            };
-
-            memory_chunk_pool_allocator() throw()
-              : memory_pool_(0)
-            {}
-            memory_chunk_pool_allocator(
-                    util::memory_chunk_pool<Mutex> & mp) throw()
-              : memory_pool_(&mp)
-            {}
-            memory_chunk_pool_allocator(
-                    memory_chunk_pool_allocator const & other) throw()
-              : memory_pool_(other.memory_pool_)
-            {}
-            template <typename U>
-            memory_chunk_pool_allocator(
-                    memory_chunk_pool_allocator<U, Mutex> const & other) throw()
-              : memory_pool_(other.memory_pool_)
-            {}
-
-            pointer address(reference x) const
-            {
-                return &x;
-            }
-
-            const_pointer address(const_reference x) const
-            {
-                return &x;
-            }
-
-            pointer allocate(size_type n, void* /*hint*/ = 0)
-            {
-                HPX_ASSERT(memory_pool_);
-                return reinterpret_cast<T*>(memory_pool_->allocate(sizeof(T) * n));
-            }
-
-            void deallocate(pointer p, size_type n)
-            {
-                HPX_ASSERT(memory_pool_);
-                memory_pool_->deallocate(reinterpret_cast<char*>(p), sizeof(T) * n);
-            }
-
-            size_type max_size() const throw()
-            {
-                return (std::numeric_limits<std::size_t>::max)() / sizeof(T);
-            }
-
-            void construct(pointer p)
-            {
-                new (p) T();
-            }
-
-            template <typename U>
-            void construct(pointer p, U && val)
-            {
-                new (p) typename util::decay<T>::type(std::forward<U>(val));
-            }
-
-            /** Destroy the object referenced by @c p. */
-            void destroy(pointer p)
-            {
-                p->~T();
-            }
-
-            bool operator==(memory_chunk_pool_allocator const & other)
-            {
-                return memory_pool_ == other.memory_pool_;
-            }
-
-            bool operator!=(memory_chunk_pool_allocator const & other)
-            {
-                return memory_pool_ != other.memory_pool_;
-            }
-
-            util::memory_chunk_pool<Mutex> * memory_pool_;
-        };
-    }
 }}
 
 #endif
